0.0.18 make set subscriptions faster for broadcaster to authorize
don't want broadcaster to have to make extra passes over
topics/globs, especially before authorization. This allows
for no passes before verifying authorization is sensible,
then 1 to verify the etag is as indicated, then 1 to set the
subscriptions.

Before, required 1 to produce the verified etag, then verify
authorization, then 1 to set the subscriptions, which is
easier to DOS